// UI-MySql/src/app/models/customer.ts
export interface Customer {
    Id: number;
    Title: string;
    CustomerName: string;
    EntityTypeId: number;
    Email: string;
    MobileNumber?: string;
    CustomerGstin?: string;
    FillingStatusId?: number;
    BranchName?: string;
    DisplayName?: string;
    PhoneNumber?: string;
    Fax?: string;
    AccountNumber?: string;
    AccountName?: string;
    BankId?: number;
    IfscCode?: string;
    AccountTypeId?: number;
    BankingBranchName?: string;
    Address1?: string;
    Address2?: string;
    City?: string;
    Pincode?: string;
    StateId?: number;
    CountryId?: number;
    AccountBranchName?: string;
    AccountGstin?: string;
    AccountEmail?: string;
    AccountMobileNumber?: string;
    ShippingAddress1?: string;
    ShippingAddress2?: string;
    ShippingCity?: string;
    ShippingPincode?: string;
    ShippingStateId?: number;
    ShippingCountryId?: number;
    ShippingBranchName?: string;
    ShippingGstin?: string;
    ShippingEmail?: string;
    ShippingMobileNumber?: string;
}